+++ Bubble.js
@@ -1,0 +1,75 @@
+import React, {Text, View, Animated, Image, StyleSheet} from 'react-native';
+
+let styles = StyleSheet.create({
+  bubble: {
+    borderRadius: 15,
+    paddingLeft: 14,
+    paddingRight: 14,
+    paddingBottom: 10,
+    paddingTop: 8,
+  },
+  text: {
+    color: '#000',
+  },
+  textLeft: {
+  },
+  textRight: {
+    color: '#fff',
+  },
+  bubbleLeft: {
+    marginRight: 70,
+    backgroundColor: '#e6e6eb',
+    alignSelf: "flex-start",
+  },
+  bubbleRight: {
+    marginLeft: 70,
+    backgroundColor: '#007aff',
+    alignSelf: "flex-end"
+  },
+  bubbleError: {
+    backgroundColor: '#e01717'
+  },
+
+});
+
+export default class Bubble extends React.Component {
+
+  constructor(props) {
+    super(props);
+  }
+
+  renderText(text = "", position) {
+
+    if (this.props.renderCustomText) {
+      return this.props.renderCustomText(rowData, rowID);
+    }
+    return (
+      <Text style={[styles.text, (position === 'left' ? styles.textLeft : styles.textRight)]}>
+        {text}
+      </Text>
+    );
+  }
+
+  render(){
+    var flexStyle = {};
+    if ( this.props.text.length > 40 ) {
+     flexStyle.flex = 1;
+    }
+
+    return (
+      <View style={[styles.bubble,
+        (this.props.position === 'left' ? styles.bubbleLeft : styles.bubbleRight),
+        (this.props.status === 'ErrorButton' ? styles.bubbleError : null),
+        flexStyle]}>
+        {this.renderText(this.props.text, this.props.position)}
+      </View>
+    )
+  }
+}
+
+Bubble.propTypes = {
+  position: React.PropTypes.oneOf(['left','rigth']),
+  status: React.PropTypes.string,
+  text: React.PropTypes.string,
+  renderCustomText: React.PropTypes.func
+}
